refactor(grid): 비차단 리뷰 권고 정리 — 쓰기 Service 경계(M2)·대형 레포 venue 추출(M5)
전체 리뷰 비차단 권고(SHIP-WITH-FIXES) 정리:
- M2 레이어링: 그리드 쓰기 변이(softTarget/slot 편집) 훅이 Repository 직결하던 것을
  gridWriteService 경유로 전환(Hook→Service→Repository 계약 준수). 동작 무변경(얇은 위임).
- M5 파일크기: venue 메서드를 sibling 모듈(JobPostingRepositoryVenue/WorkLogRepositoryVenue)로
  순수 추출(공개 시그니처/인터페이스/barrel 무변경, class 메서드 위임).
  JobPostingRepository 895→798 · WorkLogRepository 851→745 (둘 다 ≤800 하드캡).

검증(fresh): npm run quality EXIT0 · jest 37스위트/258. SQL 무변경(grid pgTAP 53/53 유효).
